requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.2
django-ckeditor==6.0.0
django-debug-toolbar==3.2.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
